test: test fd-driven tasks

+++ lib/orogen/test.rb
@@ -87,24 +87,26 @@ module Orocos
 	    end
 
             def build_test_component(dirname, with_corba, test_bin = nil)
-                # Copy +dirname+ in place of wc
-                wc_dirname = File.join(TEST_DIR, "wc")
-                FileUtils.rm_rf wc_dirname
-                FileUtils.cp_r dirname, wc_dirname
+                if !ENV['TEST_SKIP_REBUILD']
+                    # Copy +dirname+ in place of wc
+                    wc_dirname = File.join(TEST_DIR, "wc")
+                    FileUtils.rm_rf wc_dirname
+                    FileUtils.cp_r dirname, wc_dirname
 
-                in_wc do
-                    spec = Dir.glob("*.orogen").to_a.first
-                    component = Component.load(spec)
-                    if with_corba
-                        component.enable_corba
-                    else
-                        component.disable_corba
-                    end
+                    in_wc do
+                        spec = Dir.glob("*.orogen").to_a.first
+                        component = Component.load(spec)
+                        if with_corba
+                            component.enable_corba
+                        else
+                            component.disable_corba
+                        end
 
-                    compile_wc(component) do
-                        FileUtils.cp 'templates/CMakeLists.txt', 'CMakeLists.txt'
-                        File.open('CMakeLists.txt', 'a') do |io|
-                            yield(io) if block_given?
+                        compile_wc(component) do
+                            FileUtils.cp 'templates/CMakeLists.txt', 'CMakeLists.txt'
+                            File.open('CMakeLists.txt', 'a') do |io|
+                                yield(io) if block_given?
+                            end
                         end
                     end
                 end
